Renamed HttpRequest ti HttpRequestContext, added docs for ctx in fastifyOAuth.d.ts, replaced authPayload to enities, explain logic
