+++ src/domain/entities/authPayload.ts
@@ -1,0 +1,9 @@
+/**
+ * Auth payload interface, which is encoded in the access token, decoded when the token is verified
+ */
+export default interface AuthPayload {
+  /**
+   * User id
+   */
+  id: number;
+}
